Inventory item email response
EMAIL RESPONSE
A new online store has been migrating to a new software system for its inventory management. The last thing they need to do for the new system to be fully operational is complete their survey response messages.

They assigned someone with no Python experience to work on it. As you'd expect, there are now bugs in the code, and you have been asked to fix the report.